Serialization module

OpenSEFA.SerializationModule — Module

Methods to load reconciliation problems from JSON files and export solutions.

Parse models described in JSON to a ReconciliationProblem.

Serialize solutions to JSON.

OpenSEFA.SerializationModule.additional_info_keys — Constant

Keys for additional serialization infos

OpenSEFA.SerializationModule.relevant_error_keys — Constant

Keys for serialization infos if solution is INFEASIBLE

OpenSEFA.SerializationModule.load_model — Method

Load model from JSON file.

OpenSEFA.SerializationModule.load_solver — Method

Parse and load a specific solver given the data as a dictionary.

OpenSEFA.SerializationModule.load_solver — Method

Parse and load a specific solver given the data in JSON format.

OpenSEFA.SerializationModule.load_solver — Method

If the JSON model specifies a solver, use it. Otherwise pass a default solver.

OpenSEFA.SerializationModule.serialize — Function

Store the problem in JSON format.

OpenSEFA.SerializationModule.serialize — Function

Store the solution in JSON format.

OpenSEFA.SerializationModule.to_dict — Method

Extract solution data and convert to a dictionary. Add more information from the solution to be serialized as needed.