Nonlinear solver interface module

OpenSEFA.NonlinearSolveInterfaceModule — Module

Interface between data reconciliation problems and NonlinearSolve.jl: a Julia package for high-performance and differentiation-enabled nonlinear solvers (Newton methods), bracketed rootfinding (bisection, Falsi), with sparsity and Newton-Krylov support.

OpenSEFA.NonlinearSolveInterfaceModule.NonlinearSolver — Type

Solve the problem using any solver available from the NonlinearSolve.jl Julia library.

OpenSEFA.NonlinearSolveInterfaceModule.nlls! — Method

Update the residual term given current values u, adding one constraint per equation.